Nullable versions of UserData variations used in userdata.set / userdata.initialize

+++ LLServer/Models/UserData/ProfileCard.cs
@@ -4,7 +4,7 @@ public class ProfileCard
 {
     private Int128 id;
 
-    public ProfileCard(long id)
+    public ProfileCard(Int128 id)
     {
         this.id = id;
     }
